requirements.txt
pyyaml
eyed3